feat(commands): merge album commands into unified lychee:recompute-album-stats
Merge BackfillAlbumFields and RecomputeAlbumStats into single unified command.

- Make album_id argument optional for dual behavior
- Without album_id: bulk backfill mode (--dry-run, --chunk options)
- With album_id: single-album recompute mode (--sync option)
- Delete obsolete BackfillAlbumFields command and test file
- Merge test suites: 13 tests, 676 assertions passing
- Update FR-003-06 to document merged command behavior

Spec impact: Updated spec.md FR-003-02/FR-003-06, plan.md I15, tasks.md T-003-53-60, knowledge-map.md, and ADR-0003 to reflect unified command design with optional album_id parameter.
