+++ app/Console/Commands/RecomputeAlbumStats.php
@@ -21,15 +21,17 @@ class RecomputeAlbumStats extends Command
 	 * @var string
 	 */
 	protected $signature = 'lychee:recompute-album-stats
-							{album_id : The ID of the album to recompute}
-							{--sync : Run synchronously instead of dispatching a job}';
+							{album_id? : Optional album ID for single-album mode}
+							{--sync : Run synchronously (single-album mode only)}
+							{--dry-run : Preview without making changes (bulk mode only)}
+							{--chunk=1000 : Number of albums to process per batch (bulk mode only)}';
 
 	/**
 	 * The console command description.
 	 *
 	 * @var string
 	 */
-	protected $description = 'Manually recompute stats for a specific album (useful for recovery after propagation failures)';
+	protected $description = 'Recompute album stats. With album_id: recompute single album. Without album_id: bulk backfill all albums.';
 
 	/**
 	 * Execute the console command.
@@ -37,6 +39,20 @@ class RecomputeAlbumStats extends Command
 	public function handle(): int
 	{
 		$album_id = $this->argument('album_id');
+
+		// Dual behavior: with album_id = single-album mode, without = bulk mode
+		if ($album_id !== null) {
+			return $this->handleSingleAlbum($album_id);
+		} else {
+			return $this->handleBulkBackfill();
+		}
+	}
+
+	/**
+	 * Handle single-album recompute mode.
+	 */
+	private function handleSingleAlbum(string $album_id): int
+	{
 		$sync = $this->option('sync');
 
 		// Validate album exists
@@ -76,5 +92,78 @@ class RecomputeAlbumStats extends Command
 
 			return Command::SUCCESS;
 		}
+	}
+
+	/**
+	 * Handle bulk backfill mode for all albums.
+	 */
+	private function handleBulkBackfill(): int
+	{
+		$dry_run = $this->option('dry-run');
+		$chunk_size = (int) $this->option('chunk');
+
+		if ($chunk_size < 1) {
+			$this->error('Chunk size must be at least 1');
+
+			return Command::FAILURE;
+		}
+
+		$this->info('Starting album fields backfill...');
+		if ($dry_run) {
+			$this->warn('DRY RUN MODE - No changes will be made');
+		}
+
+		// Get total count
+		$total = Album::query()->count();
+		$this->info("Found {$total} albums to process");
+
+		if ($total === 0) {
+			$this->info('No albums to process');
+
+			return Command::SUCCESS;
+		}
+
+		// Process albums ordered by _lft ASC (leaf-to-root order)
+		// This ensures child albums are computed before parents
+		$bar = $this->output->createProgressBar($total);
+		$bar->start();
+
+		$processed = 0;
+
+		Album::query()
+			->orderBy('_lft', 'asc')
+			->chunk($chunk_size, function (\Illuminate\Database\Eloquent\Collection $albums) use ($dry_run, &$processed, $bar): void {
+				/** @var Album $album */
+				foreach ($albums as $album) {
+					if (!$dry_run) {
+						// Dispatch job to recompute stats for this album
+						RecomputeAlbumStatsJob::dispatch($album->id);
+					}
+
+					$processed++;
+					$bar->advance();
+
+					// Log progress at intervals
+					if ($processed % 100 === 0) {
+						$percentage = round(($processed / $bar->getMaxSteps()) * 100, 2);
+						Log::info("Backfilled {$processed}/{$bar->getMaxSteps()} albums ({$percentage}%)");
+					}
+				}
+			});
+
+		$bar->finish();
+		$this->newLine(2);
+
+		if ($dry_run) {
+			$this->info("DRY RUN: Would have dispatched {$processed} jobs");
+		} else {
+			$this->info("Dispatched {$processed} jobs to recompute album stats");
+			$this->info('Jobs will be processed by the queue worker');
+			$this->warn('Note: This operation may take some time for large galleries');
+		}
+
+		Log::info("Backfill completed: {$processed} albums processed");
+
+		return Command::SUCCESS;
 	}
 }
