+++ tests/Feature/Console/RecomputeAlbumStatsCommandTest.php
@@ -29,63 +29,22 @@ use Illuminate\Support\Facades\Queue;
 use Tests\Precomputing\Base\BasePrecomputingTest;
 
 /**
- * Test recovery command (CLI-003-02).
+ * Test unified recompute/backfill command (FR-003-06, CLI-003-02, S-003-12).
  *
- * Verifies:
- * - Command dispatches RecomputeAlbumStatsJob correctly
- * - Sync mode executes immediately
- * - Async mode queues the job
- * - Invalid album_id is handled gracefully
+ * Verifies both modes:
+ * - Single-album mode: Command recomputes stats for specific album (with album_id)
+ * - Bulk backfill mode: Command backfills all albums (without album_id)
  */
 class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 {
+	// ============================================================
+	// Single-Album Mode Tests
+	// ============================================================
+
 	/**
-	 * Test command dispatches job correctly for valid album.
+	 * Test command dispatches job correctly for valid album (async).
 	 */
 	public function testCommandDispatchesJobForValidAlbum(): void
-	{
-		$user = User::factory()->create();
-		$album = Album::factory()->as_root()->owned_by($user)->create();
-
-		$photo = Photo::factory()->owned_by($user)->create([
-			'taken_at' => new Carbon('2023-06-15'),
-		]);
-		$photo->albums()->attach($album->id);
-
-		// Initially empty
-		$this->assertEquals(0, $album->num_photos);
-
-		// Run command in sync mode
-		$this->artisan('lychee:recompute-album-stats', [
-			'album_id' => $album->id,
-			'--sync' => true,
-		])
-			->expectsOutput('Album stats recomputed successfully (sync).')
-			->assertExitCode(0);
-
-		$album->refresh();
-
-		// Should be updated
-		$this->assertEquals(1, $album->num_photos);
-		$this->assertEquals('2023-06-15', $album->min_taken_at->format('Y-m-d'));
-	}
-
-	/**
-	 * Test command handles invalid album_id gracefully.
-	 */
-	public function testCommandHandlesInvalidAlbumId(): void
-	{
-		$this->artisan('lychee:recompute-album-stats', [
-			'album_id' => 'invalid-id-that-does-not-exist',
-		])
-			->expectsOutput('Error: Album not found.')
-			->assertExitCode(1);
-	}
-
-	/**
-	 * Test async mode queues the job.
-	 */
-	public function testAsyncModeQueuesJob(): void
 	{
 		Queue::fake();
 
@@ -96,14 +55,26 @@ class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 			'album_id' => $album->id,
 			// No --sync flag, should queue
 		])
-			->expectsOutput('Album stats recomputation job dispatched.')
+			->expectsOutput('✓ Job dispatched to queue')
 			->assertExitCode(0);
 
 		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class);
 	}
 
 	/**
-	 * Test sync mode executes immediately.
+	 * Test command handles invalid album_id gracefully.
+	 */
+	public function testCommandHandlesInvalidAlbumId(): void
+	{
+		$this->artisan('lychee:recompute-album-stats', [
+			'album_id' => 'invalid-id-that-does-not-exist',
+		])
+			->expectsOutputToContain('not found')
+			->assertExitCode(1);
+	}
+
+	/**
+	 * Test sync mode executes immediately and updates album.
 	 */
 	public function testSyncModeExecutesImmediately(): void
 	{
@@ -146,7 +117,7 @@ class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 		$child->appendToNode($parent)->save();
 
 		$photo = Photo::factory()->owned_by($user)->create([
-			'taken_at' => new Carbon('2023-11-25'),
+			'taken_at' => new Carbon('2023-11-25 10:00:00'),
 		]);
 		$photo->albums()->attach($child->id);
 
@@ -160,7 +131,7 @@ class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 		$child->refresh();
 
 		$this->assertEquals(1, $child->num_photos);
-		$this->assertEquals('2023-11-25', $child->min_taken_at->format('Y-m-d'));
+		$this->assertEquals('2023-11-25', $child->min_taken_at->toDateString());
 	}
 
 	/**
@@ -172,7 +143,7 @@ class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 		$album = Album::factory()->as_root()->owned_by($user)->create();
 
 		$photo = Photo::factory()->owned_by($user)->create([
-			'taken_at' => new Carbon('2023-07-04'),
+			'taken_at' => new Carbon('2023-07-04 10:00:00'),
 		]);
 		$photo->albums()->attach($album->id);
 
@@ -194,23 +165,189 @@ class RecomputeAlbumStatsCommandTest extends BasePrecomputingTest
 
 		// Should be corrected
 		$this->assertEquals(1, $album->num_photos);
-		$this->assertEquals('2023-07-04', $album->min_taken_at->format('Y-m-d'));
+		$this->assertEquals('2023-07-04', $album->min_taken_at->toDateString());
+	}
+
+	// ============================================================
+	// Bulk Backfill Mode Tests (without album_id)
+	// ============================================================
+
+	/**
+	 * Test bulk mode backfills computed values correctly (S-003-12).
+	 */
+	public function testBulkBackfillComputesCorrectValues(): void
+	{
+		Queue::fake();
+
+		$user = User::factory()->create();
+
+		// Create albums with photos
+		$album1 = Album::factory()->as_root()->owned_by($user)->create();
+		$photo1 = Photo::factory()->owned_by($user)->create([
+			'taken_at' => new Carbon('2023-01-15'),
+		]);
+		$photo1->albums()->attach($album1->id);
+
+		$album2 = Album::factory()->as_root()->owned_by($user)->create();
+		$photo2 = Photo::factory()->owned_by($user)->create([
+			'taken_at' => new Carbon('2023-06-20'),
+		]);
+		$photo2->albums()->attach($album2->id);
+
+		// Run bulk backfill (no album_id argument)
+		$this->artisan('lychee:recompute-album-stats')
+			->expectsOutputToContain('Starting album fields backfill')
+			->expectsOutputToContain('Dispatched')
+			->assertExitCode(0);
+
+		// Verify jobs were dispatched for both albums
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 2);
 	}
 
 	/**
-	 * Test command output provides useful feedback.
+	 * Test bulk backfill is idempotent (can re-run safely).
 	 */
-	public function testCommandOutputProvidesUsefulFeedback(): void
+	public function testBulkBackfillIsIdempotent(): void
 	{
+		Queue::fake();
+
 		$user = User::factory()->create();
 		$album = Album::factory()->as_root()->owned_by($user)->create();
 
-		$this->artisan('lychee:recompute-album-stats', [
-			'album_id' => $album->id,
-			'--sync' => true,
-		])
-			->expectsOutput('Recomputing stats for album: ' . $album->id)
-			->expectsOutput('Album stats recomputed successfully (sync).')
+		$photo = Photo::factory()->owned_by($user)->create([
+			'taken_at' => new Carbon('2023-03-10'),
+		]);
+		$photo->albums()->attach($album->id);
+
+		// Run backfill first time
+		$this->artisan('lychee:recompute-album-stats')
 			->assertExitCode(0);
+
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 1);
+
+		Queue::fake(); // Reset queue
+
+		// Run backfill second time
+		$this->artisan('lychee:recompute-album-stats')
+			->assertExitCode(0);
+
+		// Should dispatch jobs again (idempotent = safe to re-run)
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 1);
+	}
+
+	/**
+	 * Test dry-run mode does not dispatch jobs.
+	 */
+	public function testDryRunDoesNotDispatchJobs(): void
+	{
+		Queue::fake();
+
+		$user = User::factory()->create();
+		$album = Album::factory()->as_root()->owned_by($user)->create();
+
+		$photo = Photo::factory()->owned_by($user)->create([
+			'taken_at' => new Carbon('2023-05-15'),
+		]);
+		$photo->albums()->attach($album->id);
+
+		// Run backfill in dry-run mode
+		$this->artisan('lychee:recompute-album-stats', ['--dry-run' => true])
+			->expectsOutputToContain('DRY RUN MODE')
+			->expectsOutputToContain('Would have dispatched')
+			->assertExitCode(0);
+
+		// No jobs should be dispatched
+		Queue::assertNothingPushed();
+	}
+
+	/**
+	 * Test chunking with custom chunk size.
+	 */
+	public function testChunkingWorksCorrectly(): void
+	{
+		Queue::fake();
+
+		$user = User::factory()->create();
+
+		// Create 5 albums
+		for ($i = 0; $i < 5; $i++) {
+			$album = Album::factory()->as_root()->owned_by($user)->create();
+			$photo = Photo::factory()->owned_by($user)->create();
+			$photo->albums()->attach($album->id);
+		}
+
+		// Run backfill with small chunk size
+		$this->artisan('lychee:recompute-album-stats', ['--chunk' => 2])
+			->assertExitCode(0);
+
+		// Verify jobs dispatched for all 5 albums
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 5);
+	}
+
+	/**
+	 * Test bulk mode handles empty album gracefully.
+	 */
+	public function testBulkBackfillHandlesEmptyAlbum(): void
+	{
+		Queue::fake();
+
+		$user = User::factory()->create();
+		$emptyAlbum = Album::factory()->as_root()->owned_by($user)->create();
+
+		$this->artisan('lychee:recompute-album-stats')
+			->assertExitCode(0);
+
+		// Job should still be dispatched for empty album
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 1);
+	}
+
+	/**
+	 * Test bulk mode processes nested albums correctly.
+	 */
+	public function testBulkBackfillProcessesNestedAlbums(): void
+	{
+		Queue::fake();
+
+		$user = User::factory()->create();
+
+		$parent = Album::factory()->as_root()->owned_by($user)->create();
+		$child = Album::factory()->owned_by($user)->create();
+		$child->appendToNode($parent)->save();
+
+		$photo = Photo::factory()->owned_by($user)->create([
+			'taken_at' => new Carbon('2023-08-20'),
+		]);
+		$photo->albums()->attach($child->id);
+
+		$this->artisan('lychee:recompute-album-stats')
+			->assertExitCode(0);
+
+		// Jobs should be dispatched for both parent and child
+		Queue::assertPushed(\App\Jobs\RecomputeAlbumStatsJob::class, 2);
+	}
+
+	/**
+	 * Test bulk mode handles no albums gracefully.
+	 */
+	public function testBulkBackfillHandlesNoAlbums(): void
+	{
+		Queue::fake();
+
+		// No albums in database
+		$this->artisan('lychee:recompute-album-stats')
+			->expectsOutputToContain('No albums to process')
+			->assertExitCode(0);
+
+		Queue::assertNothingPushed();
+	}
+
+	/**
+	 * Test invalid chunk size is rejected.
+	 */
+	public function testInvalidChunkSizeIsRejected(): void
+	{
+		$this->artisan('lychee:recompute-album-stats', ['--chunk' => 0])
+			->expectsOutputToContain('Chunk size must be at least 1')
+			->assertExitCode(1);
 	}
 }
